Reworked JS to Typescript
