+++ index.html
@@ -27,7 +27,7 @@
   <div class="gaming-window">
     <div id="main-menu" class="main-menu">
       <div class="menu">
-        <button id="save-game-button" onclick="saveGame()">Save</button>
+        <button id="save-game-button">Save</button>
       </div>
     </div>
     <div id="dialogue-scene" class="dialogue-container">
