+++ scripts/index.ts
@@ -13,6 +13,7 @@ import "./views/mainMenu"
 import "./views/dialogue"
 import "./views/town"
 import "./views/battleScene"
+import "./views/saveScreen"
 
 let currentBackground = -1;
 let playerGotStarter = false;
@@ -176,242 +177,7 @@ export const pickPokemon = (pokemonId: string) => {
 };
 
 
-let allowUserMovementInput = true;
-const playerStartX = 50;
 
-const playerMaxX = 90;
-const playerMinX = -290;
-
-const cameraMaxX = 50;
-const cameraMinX = -250;
-
-let keyADown = false;
-let keyDDown = false;
-
-let cameraLagSpeed = 0.05;
-let playerLagSpeed = 0.4;
-
-let cameraX = playerStartX;
-let cameraTargetX = playerStartX;
-
-let playerX = playerStartX;
-let playerTargetX = playerStartX;
-
-let playerSpeed = 25;
-let enteringWilderness = false;
-
-const setCameraAndPlayerProperties = () => {
-  const town = document.getElementById("town");
-  town.style.setProperty("--cameraX", `${cameraX}%`);
-  town.style.setProperty("--playerX", `${playerX}%`);
-};
-
-document.addEventListener("keydown", (e) => {
-  if (e.code === "KeyA") {
-    keyADown = true;
-  } else if (e.code === "KeyD") {
-    keyDDown = true;
-  }
-});
-
-document.addEventListener("keyup", (e) => {
-  if (e.code === "KeyA") {
-    keyADown = false;
-  } else if (e.code === "KeyD") {
-    keyDDown = false;
-  }
-});
-
-const deltaTimeSeconds = 1 / 60;
-
-console.log(deltaTimeSeconds);
-
-let playerDirection = "left";
-let playerRunning = false;
-
-const [enableTownClock, disableTownClock] = (function () {
-  let intervalId = null;
-
-  const enableTownClock = () => {
-    intervalId = setInterval(async () => {
-      if (allowUserMovementInput) {
-        if (keyADown) {
-          playerTargetX -= playerSpeed * deltaTimeSeconds;
-          playerRunning = true;
-          playerDirection = "left";
-        } else if (keyDDown) {
-          playerTargetX += playerSpeed * deltaTimeSeconds;
-          playerRunning = true;
-          playerDirection = "right";
-        } else {
-          playerRunning = false;
-        }
-
-        playerTargetX = Math.min(
-          playerMaxX,
-          Math.max(playerMinX, playerTargetX)
-        );
-        cameraTargetX = playerTargetX; // Follow player
-        cameraTargetX = Math.min(
-          cameraMaxX,
-          Math.max(cameraMinX, cameraTargetX)
-        );
-
-        if (playerTargetX + 2 > playerMaxX) {
-          if (!enteringWilderness) {
-            enterWilderness();
-            console.log("Attempting to enter wilderness.");
-            enteringWilderness = true;
-            return;
-          }
-        }
-      }
-
-      updatePlayerSprite();
-
-      cameraX = cameraX + (cameraTargetX - cameraX) * cameraLagSpeed;
-      playerX = playerX + (playerTargetX - playerX) * playerLagSpeed;
-
-      setCameraAndPlayerProperties();
-    }, deltaTimeSeconds * 1000);
-  };
-
-  const disableTownClock = () => {
-    if (intervalId) {
-      clearInterval(intervalId);
-      intervalId = null;
-    }
-  };
-
-  return [enableTownClock, disableTownClock];
-})();
-
-let spriteState = "";
-
-const updatePlayerSprite = () => {
-  const allySprite = document.getElementById("player-character") as HTMLImageElement;
-  const spriteSource = "../images/character.gif";
-  const newSpriteState = playerRunning + playerDirection;
-  const shouldUpdateSprite = spriteState != newSpriteState;
-  spriteState = newSpriteState;
-
-  if (playerRunning) {
-    if (playerDirection == "left") {
-      if (shouldUpdateSprite) {
-        allySprite.src = spriteSource;
-        allySprite.style.transform = "scaleX(1)";
-        console.log(allySprite.src);
-        // Set image to running left
-      }
-    } else {
-      if (shouldUpdateSprite) {
-        allySprite.src = `../images/character.gif`;
-        allySprite.style.transform = "scaleX(-1)";
-        // Set image to running right
-      }
-    }
-  } else {
-    if (playerDirection == "left") {
-      // Set image to idle left
-    } else {
-      // Set image to idle right
-    }
-  }
-};
-
-const enterWilderness = async () => {
-  if (playerPokemonList[0].currentHp == 0) {
-    return;
-  }
-  
-  allowUserMovementInput = false;
-  
-  await animate(
-    (time, deltaTime) => {
-      playerTargetX += playerSpeed * deltaTime;
-    },
-    2,
-    30
-  );
-  console.log("Animation done");
-  startBattle();
-  disableTownClock();
-};
-
-export const returnFromWilderness = async () => {
-  enableTownClock();
-
-  allowUserMovementInput = false;
-  playerX = playerMaxX + 15;
-  playerTargetX = playerX;
-
-  await sleep(2000);
-
-  await animate(
-    (time, deltaTime) => {
-      playerTargetX -= playerSpeed * deltaTime;
-    },
-    1.5,
-    30
-  );
-
-  enteringWilderness = false;
-  allowUserMovementInput = true;
-};
-
-let isHealing = false;
-
-
-const healPokemonTeam = async () => {
-  if (isHealing) {
-    console.log("Healing is already in progress.");
-    return; 
-  }
-
-  isHealing = true; 
-
-  playSound("pokemonshopheal.mp3");
-
-  if (!playerPokemonList) {
-    isHealing = false; 
-    return;
-  }
-
-  for (let i = 0; i < playerPokemonList.length; i++) {
-    playerPokemonList[i].currentHp = playerPokemonList[i].pokemonType.health(playerPokemonList[i].level);
-    console.log(playerPokemonList[i].currentHp);
-  }
-
-  updateVisiblePokemonInfo();
-
-  allowUserAction = false;
-  allowUserMovementInput = false;
-
-  await sleep(3000); 
-
-  allowUserMovementInput = true;
-  allowUserAction = true;
-
-  isHealing = false;
-};
-
-
-const animate = (animationFunction, durationSeconds, frameRate) => {
-  return new Promise<void>((resolve) => {
-    let time = 0;
-    let deltaTime = durationSeconds / frameRate;
-    const id = setInterval(() => {
-      if (time > durationSeconds) {
-        clearInterval(id);
-        resolve();
-        return;
-      }
-
-      time += deltaTime;
-      animationFunction(time / durationSeconds, deltaTime);
-    }, deltaTime * 1000);
-  });
-};
 
 const fetchData = () => {
   return {
@@ -426,7 +192,7 @@ const fetchData = () => {
 };
 
 const checkForLocalData = () => {
-  const loadbutton = document.getElementById("load-button");
+  const loadbutton = document.getElementById("start-screen-button-load");
   if (localStorage.getItem("saveGame") === null) {
     loadbutton.style.display = "none";
     return;
@@ -464,9 +230,10 @@ export const loadGame = () => {
 };
 
 let menuOpen = false;
-export const openMenu = (e: KeyboardEvent) => {
+export const openMenu = (e: KeyboardEvent): void => {
   const menu = document.getElementById("main-menu")
-  if (e.key == "Escape") {
+  console.log("Should open menu")
+  if (e.key === "Escape") {
     if (!menuOpen) {
       console.log("Open menu")
       menu.style.display = "block";
@@ -478,6 +245,7 @@ export const openMenu = (e: KeyboardEvent) => {
   }
 }
 
+console.log("Hej")
 document.addEventListener("keydown", openMenu);
 
 const addPokemonToTeam = (id) => {
@@ -486,6 +254,7 @@ const addPokemonToTeam = (id) => {
     createPokemonIndivual(pokemonType, 5)
   );
 }
+
